channel selection (basic) + remove openCV for matplotlib
Added the possibility to change the default channel (default is the first text channel of your server).
My goal was to setup my bot on a raspberry pi but the libraire openCV is really hard to install for this machine and I use it like two time so now I use matplotlib for image gestion.